pyproject.toml
[project]
name = "redmine-rag"
version = "0.1.0"
description = "Agent-first RAG platform for Redmine issues, journals, and wiki content"
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
  "aiosqlite>=0.21.0",
  "alembic>=1.16.5",
  "fastapi>=0.118.0",
  "httpx>=0.28.1",
  "numpy>=2.3.0",
  "orjson>=3.11.3",
  "pydantic-settings>=2.11.0",
  "sqlalchemy[asyncio]>=2.0.43",
  "tenacity>=9.1.2",
  "typer>=0.17.4",
  "uvicorn[standard]>=0.37.0",
]

[project.optional-dependencies]
ml = [
  "sentence-transformers>=5.1.1",
  "scikit-learn>=1.7.2",
]
local-llm = [
  "ollama>=0.6.0",
]

[dependency-groups]
dev = [
  "mypy>=1.18.2",
  "pre-commit>=4.3.0",
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "pytest-cov>=7.0.0",
  "pytest-dotenv>=0.5.2",
  "ruff>=0.13.1",
]

[project.scripts]
redmine-rag = "redmine_rag.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "ASYNC", "DTZ", "C4", "PIE", "PT"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q --strict-markers"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
strict_optional = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = ["alembic.*"]
ignore_missing_imports = true

[tool.alembic]
script_location = "migrations"
prepend_sys_path = [".", "src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/redmine_rag"]
